Add distribution module to Zanata server and produce a kerberos specific deployment.
The deployment process is changed to now create distribution packages
with the different deployables and necessary initial scripts. Some of
the configuration options have been moved to the database like the
authentication mechanism that the Zanata instance is configured for and
any extensions to the configured JBoss authentication modules. A new
database initialization script is now included in the distribution that
should be modified and run to add these initial application parameters.