rhbz877291 - Reset credentials after a failed login attempt.
If this is not done, a failed login will cause future login attempts (with the same user name and password) to fail. This would be ok in most cases, except when using open id authentication mechanisms, where the user name may be the same and  the password is usually empty.