Update arquillian to the most recent versions
Arquillian upgraded to 1.1.1.Final
JBoss AS managed container extension upgraded to 7.2.0.Final which brings it in line with EAP 6.1.0
Seam extension updated to 1.0.0.Beta1. This removes the need for a custom class.
We are still using a custom ManagementClient as arquillian hasn't fixed this part. This class has been updated to the same version of Arquillian.
Update the ArquillianSuiteExtension with latest changes from Aslak Knutsen.