refactor(editor): use jsf page for editor index file
This is needed because the script paths have to be accurate.
The editor has been using local script paths but they will be
disrupted when the editor is behind a redirect. It would be
bad to redirect the script requests as well since we want browsers
to cache them.
Also removes gulp since no gulp tasks are being used.