chore(editor): allow dev server to use docker server backend
Allows the editor dev server run with `make watch` to access the docker
dev server run with sercer/docker/rundev.sh.
This should remove the need for fake-zanata-server. When the editor was
a completely separate module in a different repository there was some
value in using the fake server to avoid the need to set up zanata-server,
but they now share the same repository and zanata-server can be easily
deployed with docker so the value is no longer there.