fix(editor): fix status bar rendering in standards mode
The editor was rendering in quirks mode due to an invalid DOCTYPE declaration
in the old HTML file. The new XHTML file for the editor is generating a valid
DOCTYPE declaration now.
There were some values with missing units that were treated as px in quirks
mode but were ignored completely as invalid in standards mode. Adding px units
fixed them.