Change frontend paths ZNTA-1279 (#40)
* WIP: serve frontend app from /profile/view/{username}

Client-side loads the profile properly at that URL.
 - yet to fix all the links to that path
 - changed hashHistory to browserHistory
 - all the hash URLs need to be updated to non-hash URLs

* WIP: frontend links to profile all load properly

Fixed all the routes, redirects and links for the profile view in the
main navigation of the frontend app.
 - need to make these same changes for all the other hash-based links
 - need to change any links to this in the rest of the webapp
 - need to change any links to this in other parts of the frontend app

* WIP: fixing explore up, needs more fixing

The search for people and maybe another thing are broken because hash
history is not used now. Check into the component to see what it was
doing and how to fix it.

* WIP: explore fixed

This swap of a custom history in place of the hashHistory seems to have
the search results working for explore.
 - still need to check whether I should expect to see zero users by default.

* WIP: refactor: delete temp comments

* WIP: messing with rewrite rules, and some client side stuff

* WIP: add helper method for multiple paths

* refactor: remove unused imports

* serve frontend app when glossary url is requested

* refactor(editor): use jsf page for editor index file

This is needed because the script paths have to be accurate.
The editor has been using local script paths but they will be
disrupted when the editor is behind a redirect. It would be
bad to redirect the script requests as well since we want browsers
to cache them.
Also removes gulp since no gulp tasks are being used.

* include editor config and avoid generating old editor index.html

* route editor properly client-side

* chore(editor): make dev server return app for any non-bundle URL

For paths with no trailing file extension, historyApiFallback: true would be enough,
but since the URL ends in a document name that can include extensions such as .txt,
a regular expression is needed.

* chore(editor): add config in dev index file

This config will work when running against the zanata server development
server. It may stop the fake server from working, but that is easy to
fix and is a low priority at the moment.

* refactor: adjust whitespace

* chore(editor): fix bundle paths for dev server

Bundles were being prefixed with the whole path, when they
should have just been served relative to the server.

* chore(editor): allow dev server to use docker server backend

Allows the editor dev server run with `make watch` to access the docker
dev server run with sercer/docker/rundev.sh.
This should remove the need for fake-zanata-server. When the editor was
a completely separate module in a different repository there was some
value in using the fake server to avoid the need to set up zanata-server,
but they now share the same repository and zanata-server can be easily
deployed with docker so the value is no longer there.

* SQUASH: make dev server return app for any non-bundle URL

* feat(editor): change path for editor

Changes path structure of editor. May require some further tweaks to
keep everything working as expected.

* fix(editor): fix status bar rendering in standards mode

The editor was rendering in quirks mode due to an invalid DOCTYPE declaration
in the old HTML file. The new XHTML file for the editor is generating a valid
DOCTYPE declaration now.
There were some values with missing units that were treated as px in quirks
mode but were ignored completely as invalid in standards mode. Adding px units
fixed them.

* chore(editor): make dev watch build render in standards mode

* fix(editor): fetch workspace data based on state instead of display

Was fetching workspace data when some component was mounted, but
the fetching should be based on when the workspace identity is
known rather than anything to do with the display.
Changed to observe the relevant part of state and perform the request
when that becomes available.

* fix(editor): use correct URL for document change link

Document selector was using old hash URL structure, changed to use
new URL format instead.

* fix(editor): show display name for translation locale

Was always showing locale code instead of name. This shows name when
it is available.

* feat(editor): update button in editor to point to new alpha editor location

* test(editor): update test for docs dropdown links

* fix(editor): properly handle documents with paths

* fix(editor): recommend correct path structure when wrong editor path is used

* refactor(AppView): add explanatory comment for leaving docId unencoded.

* refactor(UrlRewrite): remove unused comments

* refactor(UrlRewrite): remove unused rules (effectively dead code)

* fix(frontend): move file that was missed in migration

* fix(frontend): correct git mirgration error

Somehow during the repository migration process, a change to RoutingHelpers.js
in frontend was transposed to instead be a change in RoutingHelpers.js in editor.
The rewritten commit has the wrong file changed.

* fix(ZNTA-1478): change languages path to use non-hash URL

* fix(ZNTA-1480): change URL from /a/more to /info

* fix(whitelist): move editor dev access whitelist to the correct file

Was mistakenly put in overlay during a migration or rebase. Moved to the
docker cli script since that is now the appropriate place to set system
properties for the docker dev container.

* fix(Nav): prevent duplicate context path in navigation links

* Merge branch 'master' into damason/retry-rebase

Conflicts:
  server/zanata-frontend/pom.xml
    Several changes in master. This branch just has the processhtml
    section removed and added a property to skip checkstyle on the
    javascript code. Kept it identical to master in all other changes.