fix(ZNTA-61): prevent invalid DOM nesting with ListGroupItem
This is a bug or design flaw in react-bootstrap where it will wrap the
contents of a ListGroupItem in a <p> tag when a header is specified. It
should use a <div> instead so that panels can be used for any content.
The workaround is to specify the header as a child manually so that the
wrapping is not triggered.