fix(ZNTA-1555): link atomic css stylesheet separately
The atomic css generation utilities appear to be designed to output a
separate file that does not go through any other pipelines. It is far
easier to just link it as a separate file than to try to get it bundled
into the frontend.css file. This should be fine for the moment; I
consider it less of an annoyance than having to run an extra build to
generate the file to then import it on the next build.