chore(frontend dev builds): fix rebuild for frontend and editor watch tasks
Something was going wrong with rebuild, even though the first build worked.
It seems like it was related to multiple-modules not playing well with
webpack-dev-server, so making separate configs with a single entrypoint each
did the trick.
Hot reload is still not working quite as expected - a refresh is still
needed - but this probably needs some more modules and config to make it
work. Better do that separately rather than complicate this fix.