feat(ZNTA-1735): show icons in storybook without duplicating icon svg
Including an <Icons /> component more than once in the component tree
will insert multiple copies of the icon svg onto the page. This will
bloat the memory use of the page, and could impact performance in
other ways.
The correct approach at the moment is to use .addDecorator in each
story.js file (for components that use any icons).