fix(glossary): defactor glossary to use binding inline in JSX to satisfy react-list
The old version of react-list we are using seems to break unless we bind the
update function inline - worst-practice but it just does not work when binding
in the constructor or let React bind it by using an arrow function.

The render function does not appear to be part of the newer API for react-list,
it takes a child component as a template instead. Updating to the new version
may help bring us in line with good practice.