Prep for React upgrade - remove createClass (ZNTA-1953) (#328)
* refactor(frontend): convert frontend components to use classes

This is preparation to update React to v15

* fix(frontend): upgrade react-chartjs to remove a deprecation warning

* fix(ZNTA-1935): remove warnings and prevent React swallowing props.key

React silently consumes props.key and it is not defined in the component
this.props. The workaround is to use a different name (e.g. id). We
should look at adding a linting rule to make using 'key' as a prop
name in propTypes an error.

* refactor(frontend): stop using manual binding on React component methods

Using arrow functions when defining a custom method on a component
allows the method to be passed as a callback without the need to
manually bind it. Custom methods are any that are not React lifecycle
methods.

* refactor(ZNTA-1953): use static field for component propTypes and arrow functions for custom methods

This is to make all class-based components consistent in how they
specify propTypes and non-lifecycle methods.

* fix: minor linting errors

* fix(glossary): defactor glossary to use binding inline in JSX to satisfy react-list

The old version of react-list we are using seems to break unless we bind the
update function inline - worst-practice but it just does not work when binding
in the constructor or let React bind it by using an arrow function.

The render function does not appear to be part of the newer API for react-list,
it takes a child component as a template instead. Updating to the new version
may help bring us in line with good practice.