chore(frontend watch): change from global config to config component
This replaces access to window.config with an config component that
can be mocked. This will allow much easier changes to how we pass
config into the app.
This also splits the base URL config into 3 separate URLs to allow
finer control, such as allowing search and navigation in the watch
mode (`make watch`) by passing in the relevant URLs.