chore(ZNTA-1922): upgrade react-chartjs to be compatible with react@15.5.x
Note that this is not an official release. The official release seems
stagnant and has not been updated to work without errors in react 15.5.
This points to a commit based on the most recent release but with
a minimal change that fixes the warnings.
I will likely change to use react-chartjs-v2 since it is better maintained,
but this may require some changes to the structure of the chart data so I
will defer it for now.