Upgrade to React 15.5 (ZNTA-1922) (#397)
* chore(ZNTA-1922): upgrade react to v15.5.x

This introduces numerous warnings that will be fixed in subsequent commits.

* chore(ZNTA-1922): upgrade/remove peer dependencies of react@15.5.x

Some peer dependencies are moved to other modules. The imports will be
updated in the following commits.

* chore(ZNTA-1922): replace react-addons-test-utils with react-dom/test-utils

This is the same module, it was just moved into the react-dom package.

* chore(ZNTA-1922): replace react-addons-update with immutability-helper

As described in React 15 blog post:
https://facebook.github.io/react/blog/2017/04/07/react-v15.5.0.html

* chore(ZNTA-1922): replace React.PropTypes with prop-types module

Using React.PropTypes is deprecated, so the separate module is now
the standard way to use PropTypes.

* chore(ZNTA-1922): replace redux-simple-router with react-router-redux

They are the same package, just renamed to react-router-redux when it
went from version 2 to 3.

* chore(ZNTA-1922): upgrade react-router and react-router-redux to be compatible with react@15.5.x

Updated to versions that are compatible with React 15.
There is a newer major version of react-router, but the compatible
react-router-redux for that is still in beta.

* chore(ZNTA-1922): upgrade react-redux to latest

Major version upgrade but they assert it is API compatible.

* chore(ZNTA-1922): fix frontend routing with simplistic fix

This does a quick-fix using locationBeforeTransitions, which may just
be the same as the old state.routing.location but with a more meaningful
name, but that is not clear and this may need further work.
The react-router-redux middleware may be a good fit.

* chore(ZNTA-1922): fix editor routing

* chore(ZNTA-1922): upgrade react-intl to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-bootstrap and react-overlays to be compatible with react@15.5.x

Were causing deprecation warnings for React 15.5

* chore(ZNTA-1922): use prop-types in EditableText

Changed most of them earlier, but missed this one.

* chore(ZNTA-1922): upgrade react-textarea-autosize to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-loading to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-select to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-split-pane to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-sidebar to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-helmet to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-list to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-autosuggest to be compatible with react@15.5

* chore(ZNTA-1922): upgrade react-date-range to be compatible with react@15.5.x

* chore(ZNTA-1922): upgrade react-chartjs to be compatible with react@15.5.x

Note that this is not an official release. The official release seems
stagnant and has not been updated to work without errors in react 15.5.
This points to a commit based on the most recent release but with
a minimal change that fixes the warnings.
I will likely change to use react-chartjs-v2 since it is better maintained,
but this may require some changes to the structure of the chart data so I
will defer it for now.

* fix(ZNTA-1922): correct typo in prop name

Was bStyle instead of bsStyle. Also fixes some jsx sections that were
not wrapped to expand variables.

* fix(ZNTA-1922): remove unrecognized tag left over from different element type

* fix(ZNTA-1922): use correct link type in language teaser

* fix(ZNTA-1922): remove and replace unknown props

These look like they are generally from changing component type and
not checking that the props still make sense. We need a test suite that
exercises all the components to catch this type of error when devs
miss it.

* fix(ZNTA-1922): correct elements that were improperly migrated from <Flex>

These look like improperly migrated <Flex> elements.

* fix(ZNTA-1922): stop incorrect use of some react-bootstrap components

* fix(ZNTA-1922): use debounce correctly for search on language page

It was never invoking the search, and was not persisting the debounced
callback so even if it were called it would not debounce properly.
This gist shows what was wrong and how the fix works:
https://gist.github.com/davidmason/2ad6fc4b35209a4fb6778cf3deeaead7

* fix(ZNTA-1922): focus language id search input when it is clicked

Clicking on the right of the field was not focusing the input field
properly. This ensures the input element is the full width so it will
focus with a click anywhere in the field.

* fix(ZNTA-1922): stop changing between controlled and uncontrolled components

This was happening when component state was reset to be undefined, which
was passed as the input value. An undefined input value defines an input
as uncontrolled.

* fix(ZNTA-1922): remove attribute accidentally copied over from angular code

This happened quite some time ago, but I was able to trace it back
to the source. I do not think it has ever worked in the React app
so there is no harm in removing it.

* fix(ZNTA-1922): avoid extra padding between content and sidebar

This was introduced by a change in react-sidebar:
https://github.com/balloob/react-sidebar/commit/8f2e24092ba70625af6901cf14074869e3daef63

* refactor: minor whitespace changes

* chore(ZNTA-1922): replace react-chartjs with react-chartjs-2

* chore(ZNTA-1922): return contribution chart to its former glory

It is not identical, but as close as the new version of the charting library
seems to be capable of.

* fix(ZNTA-1922): fix invalid table structure in glossary view

This was probably caused by an update to react-bootstrap.

* fix(contribution chart): lock chart scale between 0 and 10 when contributions are low

This prevents showing negative numbers on the scale when contributions are all 0, and
ensures that very low contribution numbers will show as a low graph.

* fix(console warning): prevent warning for missing callback

The componnet does not always need it, but marks it as .isRequired.
The proper fix here will be to split the component into 2 components
since they need a very different set of props input.