chore(ZNTA-2119): add initial component snapshots
Generated by storyshots.
Note that these are large files with snapshots for all the components
that have stories. This is not ideal, but there is a pull request in
progress for storyshots that will change this so that snapshots are
stored in the component directory, which will make more sense.
See: https://github.com/storybooks/storybook/pull/1584