Activate storybook addons (ZNTA-2119) (#478)
This activates some storybook addons:

 - host: allows sizing and positioning of the storybook container to better present the component
 - info: shows story code, props details and optional notes
 - knobs: allows changing story input values in a story
 - storyshots: renders each storybook as a snapshot test, which produces a test failure when components render different markup.

Squashed merge of the following:

* chore(ZNTA-2119): add storybook info addon

Includes example usage on editor <Button> component.

* chore(ZNTA-2119): add storybook host addon

Host addon allows positioning the element and providing a specific
size of context to display the component in, as well as a few other
decorations.

* chore(ZNTA-2119): add storybook knobs addon

Includes a button builder story that demonstrates use of the knobs
addon.

* chore(ZNTA-2119): add storyshots

This required several changes to jest and storybook config to have
the storyshots tests execute without errors.
It also includes some cleanup in some of the storybooks and components
that were generating errors or warnings in tests.

* chore(ZNTA-2119): add initial component snapshots

Generated by storyshots.
Note that these are large files with snapshots for all the components
that have stories. This is not ideal, but there is a pull request in
progress for storyshots that will change this so that snapshots are
stored in the component directory, which will make more sense.
See: https://github.com/storybooks/storybook/pull/1584

* chore(ZNTA-2119): clean up jest mocks

* feat(ZNTA-2119): remove commented-out lines

* feat(ZNTA-2119): fix test failure due to missing TriCheckbox snapshots

Snapshots cannot be written in CI builds, so missing snapshots will fail tests.