Just warn if lock files are found on NFS
This avoids the need to manually delete lock files, which is nearly
always too paranoid for a non-clustered application. Until we set up
clustering, it should be safe to delete the lock files, even on NFS, as
long as there are other ways of avoiding multiple instances of Zanata
(eg some sort of native process/service management). But we log a
warning just in case.