refactor: fix up some commented-out code saved during conflict resolution
There were a few changes to a file in master, but the contents of that file
were extracted to a separate module in my branch. Since they were just
commented-out code as a reminder when activating a feature, I just kept all
the added parts at the end of their original file and manually move them
to the right place in this commit.