chore(flow): add flow type definitions for our libraries
These type definitions allow Flow to check types that we pass to and get from
our third-party libraries.
These are installed with the command-line utility flow-typed. I have chosen to
check in these type definitions because flow-typed is a bit flaky and cna take
up to a few tens of minutes to run so it is not practical to have it run
as part of a build or have each developer run it individually.