refactor(actions): remove unused action FETCH_FAILED
The action was not handled by any reducer. It was only used in places where
we should be using api middleware and handling errors in the standard way,
so I have just put some comments in the code and will handle that soon when
updating all the code to use api middleware.