Add settings component with a few settings (ZNTA-1927) (#524)
Adds a user settings API endpoint, and makes the React editor load settings from that endpoint on startup. It adds an implicit setting for whether the suggestions panel is open, and an explicit setting to change the 'Enter' key behaviour.

* feat(ZNTA-1927): update docker instructions to clarify how to find port

* feat(ZNTA-1927): prefix gwt editor settings keys with "gwt."

* feat(ZNTA-1927): add options REST endpoints

Includes GET and POST endpoints, which include a prefix for the option names
to return.

* test: update jest to 21.x and update a few snapshots

This version of Jest is needed to properly compare objects that use symbol
keys. In this branch the API call action tests require this.
8 snapshots were broken due to different pretty-print indentation levels
in the new version of Jest (no real differences).

* feat(ZNTA-1927): add user settings actions and reducer

* feat(ZNTA-1927): use user settings for suggestion panel visibility

* test(ZNTA-1927): test settings actions and reducer

* refactor: fix typo in doc comment

* refactor: fix linting errors in SettingOption and SettingsOptions files

* feat(ZNTA-1927): add settings panel with enter key setting option

Also enables Flow for type-checking (since there were several changes
that were a pain to keep track of without it), and refactors some of
the files I needed to work with - closer to the ideal structure the
app should be using.

* fix(editor): prevent error when GlossarySearchInput is not yet bound

This was a subtle bug, triggered when switching between different
fix(editor): prevent error when GlossarySearchInput is not yet bound

* refactor: fix up some commented-out code saved during conflict resolution

There were a few changes to a file in master, but the contents of that file
were extracted to a separate module in my branch. Since they were just
commented-out code as a reminder when activating a feature, I just kept all
the added parts at the end of their original file and manually move them
to the right place in this commit.

* test(ZNTA-1927): update tests, add enzyme

Updated tests. Was having issues trying to render a stateless component
since that is not well supported by the standard TestUtils. Used enzyme
in that test instead, which supports stateless components and has a
cleaner API anyway. We should use enzyme in all our tests.

* test(ZNTA-1927): update snapshots for modals

This is just an indentation mismatch on the example background content.
We should insert the background content as a separate module so that
it can be mocked away in the snapshot tests.

* refactor(ZNTA-1927): rename SidebarSettings to SettingsPanel

* fix(ZNTA-1927): fix errors introduced during merge

Duplicate logger definition, and missing closing tag in a changelog.

* feat(ZNTA-1927): fix indentation and use stream in UserService

* refactor: remove unused comments

* feat(ZNTA-1927): remove flow-typed annotations

These will be submitted in a separate pull request

* fix: remove unused serialVersionUID and logger

* fix: remove unused import

* fix: remove some unused serialVersionUID

These are generating Java warnings

* fix(ZNTA-1927): fix compile error

* refactor(ZNTA-1927): remove unused imports and update warning threshold

* feat(ZNTA-1927): make Shift+Enter add a line when Enter saves immediately